add underline for ArtistNavbar
